requirement.txt
streamlit
pandas
numpy
opencv-python
pymongo
python-dotenv
ultralytics
torch
torchvision
torchaudio
